+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Manager/SK2ReceiptManager.swift
@@ -151,6 +151,12 @@ actor SK2ReceiptManager: ReceiptManagerType {
 
   @available(iOS 17.2, macOS 14.2, tvOS 17.2, watchOS 10.2, visionOS 1.1, *)
   private func updatePeriodType(from transaction: Transaction) {
+    #if compiler(>=6.0.0)
+    if transaction.offer?.type == .winBack {
+      latestSubscriptionPeriodType = .winback
+      return
+    }
+    #endif
     switch transaction.offer?.type {
     case .introductory:
       latestSubscriptionPeriodType = .trial
@@ -158,8 +164,6 @@ actor SK2ReceiptManager: ReceiptManagerType {
       latestSubscriptionPeriodType = .code
     case .promotional:
       latestSubscriptionPeriodType = .promotional
-    case .winBack:
-      latestSubscriptionPeriodType = .winback
     case .none:
       latestSubscriptionPeriodType = .subscription
     default: